Refactor TableFinishOperator and TableWriterOperator
- Rename methods
- Extract common fields and some utility methods into an utility class
