+++ presto-main/src/main/java/com/facebook/presto/operator/TableWriterOperator.java
@@ -44,6 +44,8 @@ import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.atomic.AtomicLong;
 
 import static com.facebook.presto.SystemSessionProperties.isStatisticsCpuTimerEnabled;
+import static com.facebook.presto.operator.TableWriterUtils.STATS_START_CHANNEL;
+import static com.facebook.presto.operator.TableWriterUtils.createStatisticsPage;
 import static com.facebook.presto.spi.type.BigintType.BIGINT;
 import static com.facebook.presto.spi.type.VarbinaryType.VARBINARY;
 import static com.facebook.presto.sql.planner.plan.TableWriterNode.CreateHandle;
@@ -62,11 +64,6 @@ import static java.util.concurrent.TimeUnit.NANOSECONDS;
 public class TableWriterOperator
         implements Operator
 {
-    public static final int ROW_COUNT_CHANNEL = 0;
-    public static final int FRAGMENT_CHANNEL = 1;
-    public static final int CONTEXT_CHANNEL = 2;
-    public static final int STATS_START_CHANNEL = 3;
-
     public static class TableWriterOperatorFactory
             implements OperatorFactory
     {
@@ -284,7 +281,7 @@ public class TableWriterOperator
             if (aggregationOutput == null) {
                 return null;
             }
-            return createStatisticsPage(aggregationOutput);
+            return createStatisticsPage(types, aggregationOutput, createTableCommitContext(false));
         }
 
         if (state != State.FINISHING) {
@@ -304,35 +301,6 @@ public class TableWriterOperator
         }
 
         state = State.FINISHED;
-        return new Page(positionCount, outputBlocks);
-    }
-
-    // Statistics page layout:
-    //
-    // row     fragments     context     stats1     stats2 ...
-    // null       null          X          X          X
-    // null       null          X          X          X
-    // null       null          X          X          X
-    // null       null          X          X          X
-    // ...
-    private Page createStatisticsPage(Page aggregationOutput)
-    {
-        int positionCount = aggregationOutput.getPositionCount();
-        Block[] outputBlocks = new Block[types.size()];
-        for (int channel = 0; channel < types.size(); channel++) {
-            if (channel < STATS_START_CHANNEL) {
-                // Include table commit context into statistics page to allow TableFinishOperator publish correct statistics for recoverable grouped execution.
-                if (channel == CONTEXT_CHANNEL) {
-                    outputBlocks[channel] = RunLengthEncodedBlock.create(types.get(channel), getTableCommitContext(false), positionCount);
-                }
-                else {
-                    outputBlocks[channel] = RunLengthEncodedBlock.create(types.get(channel), null, positionCount);
-                }
-            }
-            else {
-                outputBlocks[channel] = aggregationOutput.getBlock(channel - STATS_START_CHANNEL);
-            }
-        }
         return new Page(positionCount, outputBlocks);
     }
 
@@ -366,10 +334,10 @@ public class TableWriterOperator
             VARBINARY.writeSlice(fragmentBuilder, fragment);
         }
 
-        return new Page(positionCount, rowsBuilder.build(), fragmentBuilder.build(), RunLengthEncodedBlock.create(VARBINARY, getTableCommitContext(true), positionCount));
+        return new Page(positionCount, rowsBuilder.build(), fragmentBuilder.build(), RunLengthEncodedBlock.create(VARBINARY, createTableCommitContext(true), positionCount));
     }
 
-    private Slice getTableCommitContext(boolean lastPage)
+    private Slice createTableCommitContext(boolean lastPage)
     {
         TaskId taskId = operatorContext.getDriverContext().getPipelineContext().getTaskId();
         return wrappedBuffer(tableCommitContextCodec.toJsonBytes(
